move KirkleesBatleyEastMixin to own file
